+++ dune/common/typetraits.hh
@@ -417,6 +417,8 @@ namespace Dune
    *
    * \tparam T Variadic type list
    *
+   * \deprecated This class is deprecated and will be removed after Dune 2.10. Use sizeof...(T) instead.
+   *
    * The ::value member gives the size of the variadic type list T...
    * This should be equivalent to sizeof...(T). However, with clang
    * the latter may produce wrong results if used in template aliases
@@ -426,7 +428,9 @@ namespace Dune
    * in template aliases for any code that should work with clang < 3.8.
    */
   template<typename... T>
-  struct SizeOf
+  struct
+  [[deprecated("This class is deprecated and will be removed after Dune 2.10. Use sizeof...(T) instead.")]]
+  SizeOf
     : public std::integral_constant<std::size_t,sizeof...(T)>
   {};