+++ CHANGELOG.md
@@ -115,6 +115,7 @@ In order to build the DUNE core modules you need at least the following software
 - Deprecated the file `dune/common/assertandreturn.hh` and the contained utility
   `DUNE_ASSERT_AND_RETURN`. Use `assert()` macro directly in `constexpr` functions.
 - Remove deprecated header `power.hh`. Use `Dune::power` from `math.hh` instead.
+- Deprecate class `SizeOf`. Use `sizeof...` instead.
 
 
 # Release 2.9